requirements.txt
gradio
pandas
matplotlib
pillow
